intellij code cleanup

+++ src/main/java/com/hbm/items/food/ItemHazardSoup.java
@@ -77,7 +77,7 @@ public class ItemHazardSoup extends ItemSoup implements IItemHazard {
 			bf.posX = player.posX;
 			bf.posY = player.posX;
 			bf.posZ = player.posZ;
-			bf.destructionRange = (int) 25;
+			bf.destructionRange = 25;
 			worldIn.spawnEntity(bf);
 			if(BombConfig.enableNukeClouds) {
 				EntityNukeTorex.statFac(worldIn, player.posX, player.posY, player.posZ, 25);
@@ -92,7 +92,7 @@ public class ItemHazardSoup extends ItemSoup implements IItemHazard {
 			bf.posX = player.posX;
 			bf.posY = player.posX;
 			bf.posZ = player.posZ;
-			bf.destructionRange = (int) 50;
+			bf.destructionRange = 50;
 			worldIn.spawnEntity(bf);
 			if(BombConfig.enableNukeClouds) {
 				EntityNukeTorex.statFac(worldIn, player.posX, player.posY, player.posZ, 50);
